a bunch of refactors
