+++ client/src/utils/PostUtils.js
@@ -1,7 +1,7 @@
 const baseUrl = import.meta.env.VITE_API_URL || "http://localhost:3000";
 
 import axios from "axios";
-import { tokenize } from "./recommendationUtils.js";
+import { tokenize } from "./clientPostRecommendationUtils.js";
 import { LIKE, UNLIKE, RANKING_MODES } from "./constants.js";
 
 // Uploads a post, starting with the file attachment to GCS and then the full post data to server
